+++ src/commands/unmute.ts
@@ -160,27 +160,31 @@ async function run(
 
   const embed = new CustomEmbed(message.member);
 
-  let msg = punishAlts ? `unmuting account and any alts...` : `✅ \`${target.user.username}\` has been unmuted`;
+  const ids = await getAllGroupAccountIds(message.guild, target.user.id);
+
+  let msg = punishAlts && ids.length > 3 ? `unmuting account and any alts...` : `✅ \`${target.user.username}\` has been unmuted`;
   
   embed.setDescription(msg);
   
   let res;
 
-  if (args.join(" ").includes("-s")) {
-    if (message instanceof Message) {
-      await message.delete();
-      res = await message.member.send({ embeds: [embed] }).catch(() => {});
+  if (ids.length > 3) {
+    if (args.join(" ").includes("-s")) {
+      if (message instanceof Message) {
+        await message.delete();
+        res = await message.member.send({ embeds: [embed] }).catch(() => {});
+      } else {
+        res = await message.reply({ embeds: [embed], ephemeral: true });
+      }
     } else {
-      res = await message.reply({ embeds: [embed], ephemeral: true });
+      res = await send({ embeds: [embed] });
     }
-  } else {
-    res = await send({ embeds: [embed] });
   }
 
   let altsUnmted = 0;
 
   if (punishAlts) {
-    for (const id of await getAllGroupAccountIds(message.guild, target.user.id)) {
+    for (const id of ids) {
       if (id == target.user.id) continue;
       const unmuted = await doUnmute(
         message,
@@ -202,10 +206,23 @@ async function run(
 
   embed.setDescription(msg);
   
-  if (message instanceof Message) {
-    await (res as Message).edit({ embeds: [embed] });
+  if (ids.length > 3) {
+    if (message instanceof Message) {
+      await (res as Message).edit({ embeds: [embed] });
+    } else {
+      await message.editReply({ embeds: [embed] })
+    }
   } else {
-    await message.editReply({ embeds: [embed] })
+    if (args.join(" ").includes("-s")) {
+      if (message instanceof Message) {
+        await message.delete();
+        await message.member.send({ embeds: [embed] }).catch(() => {});
+      } else {
+        await message.reply({ embeds: [embed], ephemeral: true });
+      }
+    } else {
+     await send({ embeds: [embed] });
+    }
   }
 }
